refactor(graph): extract graph_parallel context managers out of lang/misc.py
Moves the `graph_parallel_context()` / `graph_parallel()` no-op context-manager stubs into their own
`quadrants/lang/graph_parallel.py`, mirroring `checkpoint.py`, to keep the 958-line catch-all `lang/misc.py`
(imported by 33+ modules) from growing further. Re-exported via `misc.py` so the canonical
`qd.graph_parallel_context` / `qd.graph_parallel` import paths are unchanged.

Addresses the "Check feature factorization" CI finding on PR #756.
